# Startup speed improvements