--- a/src/Files.App/Properties/PublishProfiles/win10-x64.pubxml
+++ b/src/Files.App/Properties/PublishProfiles/win10-x64.pubxml
@@ -12,6 +12,7 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
     <PublishSingleFile>False</PublishSingleFile>
     <PublishReadyToRun Condition="'$(Configuration)' == 'Debug'">False</PublishReadyToRun>
     <PublishReadyToRun Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRun>
+    <PublishReadyToRunComposite Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRunComposite>
    <!-- 
     See https://github.com/microsoft/CsWinRT/issues/373
     <PublishTrimmed>True</PublishTrimmed>