**Resolved / Related Issues**
- Avoid creating ResourceManager to get RTL settings again and again on startup
- Enable crossgen2 composite mode for more pre-generated native code optimizations.

**Validation**
How did you test these changes?
- [x] Did you build the app and test your changes?


https://user-images.githubusercontent.com/14960345/235616258-7c271e28-3cf2-43a3-8a6f-66f1621f2857.mp4

|  | Show the initial window with Mica effect | Load all things | 
| -- | -- | -- |
| Before | 2500ms | 4240ms |
| After | 1960ms | 3740ms |